master/considering/master/requirements.txt
apparmor==2.13.4
appdirs==1.4.3
application-utility==1.3.2
asgiref==3.2.5
bauh==0.9.0
btrfsutil==1.2.0
CacheControl==0.12.6
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0.post1
cupshelpers==1.0
decorator==4.4.2
distlib==0.3.0
distro==1.5.0
Django==3.0.4
django-discover-runner==1.0
django-role-permissions==2.2.1
docopt==0.6.2
fdfs-client-py==1.2.6
html5lib==1.0.1
idna==2.9
keyutils==0.6
LibAppArmor==2.13.4
louis==3.13.0
msgpack==1.0.0
mutagen==1.44.0
mysqlclient==1.4.6
npyscreen==4.10.5
ordered-set==3.1.1
packaging==20.3
pacman-mirrors==4.16.4
pep517==0.8.2
Pillow==7.0.0
powerline-status==2.7
progress==1.5
psutil==5.7.0
pwquality==1.4.2
pyaml==20.4.0
pycairo==1.19.1
pycups==2.0.1
pycurl==7.43.0.5
PyGObject==3.36.1
PyMySQL==0.9.3
pyparsing==2.4.7
PyQt5==5.14.2
PyQt5-sip==12.7.2
pysmbc==1.0.20
pyte==0.8.0
python-libtorrent==1.2.6
pytoml==0.1.21
pytz==2019.3
PyYAML==5.3.1
rados==2.0.0
rbd==2.0.0
reportlab==3.5.42
requests==2.23.0
retrying==1.3.3
rgw==2.0.0
sip==4.19.22
six==1.14.0
sqlparse==0.3.1
team==1.0
thefuck==3.30
toml==0.10.0
udiskie==2.1.1
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
